+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support for option labels in Variable-plugin
 - Generic rdfa variables input plugin - HTML only input
 
+### Added
+- Add label to variables
+
 ## [7.1.0] - 2023-05-18
 
 ### Added